Correccion de la tabla y sus columnas correspondientes

+++ src/main/java/com/project/mac/repository/IInsumoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IInsumoRepository extends JpaRepository <Insumos, Long> {
+public interface IInsumoRepository extends JpaRepository <Insumos, Integer> {
 
 
 }
